♻ refactor: utils 디렉토리 이동
